"Added new projects and updated solution file structure"
This commit adds several new projects to the solution, including 'Catalog', 'Services', 'Identity', 'Sales', 'Pricing', 'Inventory', 'Presentation', 'Manager', 'Web', 'Common', and 'Domain' within both 'Inventory' and 'Common'. Update was needed to structure the solution in a modular way, making it easier to manage and scale. Each new project is separate with its business concern which contributes to the maintainability of the whole solution.