Add Budget and BudgetItem entities to Sales domain
Added Budget and BudgetItem entities to the Sales domain. A Budget consists of a title, description, validity date, status, and a list of BudgetItems. To track products in the budget, budget items specific to each product with its quantity are included. A method for adding new budget items to the budget is also implemented. This allows a better organization and tracking of sales budgets.